Deleted adminPanel Component, added admin-panel view. Some changes on tailwind
